Basic working WMCO
The WMCO can watch WMC object called instance and
react to the changes to it. It spins up and down
Windows VMs in existing cloud provider and
makes them as Windows worker nodes in the
cluster. Also, added a e2e test suite
which validates the creation and deletion of the Windows VMs.