[build] Account for GOFLAGS not being set
"set -o nounset" results in errors on trying to access a variable which
is not set. This is an issue with GOFLAGS as it is set in the CI build
container and not in other places. Workaround this by using an
intermediate variable that is set to an empty string if GOFLAGS is not
set and use that variable to check for "-mod=vendor".