Use correct RBAC permissions
This commit changes the deploy yamls to use the correct RBAC
permissions. A big change in this commit is that the operator is now
cluster scoped. This is required as we need to work with node
objects, which requires ClusterRole permissions.

I followed
https://github.com/operator-framework/operator-sdk/blob/master/doc/operator-scope.md#changes-required-for-a-cluster-scoped-operator
to convert this to a cluster scoped operator.