[wmco] Create tracker configmap for Windows VMs
The operator needs an internal tracking mechanism for Windows
nodes that it is managing that has some associated meta data.
This meta data will allow the Operator to sync the nodes on
cluster to what is expected as per the WMC CR. This commits
add the tracker configmap which can be used to identify the
nodes that are being added to the cluster.

Co-authored-by: Aravindh Puthiyaparambil <aravindh@redhat.com>