Remove glob copying
Docker has a longstanding bug using globs
in COPY directives (i.e. COPY foo/dir/* /tmp/dir/ ).
If the source files matched by the glob include
subdirectories, these subdirectories are
not present in the destination. While this is not
a problem for WMCO Dockerfiles, we are doing it
just to be on safe side.

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1813508