[ci] React to hybrid overlay changes upstream
Because of this [commit](https://github.com/openshift/ovn-kubernetes/commit/2563e6b82827726cd7850ed2824346a6ae5ba5eb#diff-86eeef0b956d450bd2a787067b7443fdR26) in ovn-kubernetes repo, the
hybrid-overlay binary is no longer called hybrid-overlay.exe
but it is called hybrid-overlay-node.exe. We've a hardcoding
to hybrid-overlay.exe in the following locations:

 -   Dockerfile
 -   Dockefile.ci
 -   wellknownlocations.go

And there are other places where it is mentioned
in code comments.The name of this binary needs
to be changed for the operator to start working.
Otherwise the CI will keep failing.

In addition,
- [commit](https://github.com/openshift/ovn-kubernetes/commit/fc16768b4b8274a4a6d051ce85ac4161a04eda42#diff-7374600df6a5d7b56af840179f7ad27fR86)
  brought in changes related to renaming OVN overlay networks. This commit
  address that as well