[wmco] Reorder check for HNS networks
Problem: Any vm.Run() call after the hybrid overlay configuration has
completed was taking more than 5 minutes to complete. This is causing
the tracker validation test to be flaky as its timeout was 10 munutes.

Solution: To mitigate this
openshift/windows-machine-config-bootstrapper#182 introduces
reinitializing the WinRM client as part of the Reinitialize() call.
Reorder the waitForHNSNetworks() call after the Reinitialize() call to
take advantage of this.
We are adding a 2 minute sleep after the hybrid overlay starts running
to allow it complete its work.

This fix has the very useful side effect of making the node preparation
time lower.