[wmco] Generate cluster address from config
As of now, we're relying on an environment variable for
cluster address. If the operator runs in a container,
we are mounting the cluster address as the secret
to be used within operator. This commit removes
the dependency on the environment variable. We can
now generate the cluster address from apiserver
config.

https://issues.redhat.com/browse/WINC-274