[wmco] fix validation for replicas
fixes validation for replicas field as it was being interpreted as
float while validating.
this was because `replicas` was defined as `int` in place of `int32`
which is causing the `operator-sdk generate crds` to not create a
`format` field in crd. This is a known issue in
operator-sdk 0.17.

revert back this change once we bump operator-sdk version to
0.18 or above

Bug 1835166