[wmco] Reconstruct WindowsVM map from the tracker cm
As of now, if the WMCO is restarted, the knowledge of the nodes
created via WMCO is lost. We can construct this information from
the tracker configmap. This commit addresses the following items:

- https://issues.redhat.com/browse/WINC-269
- https://issues.redhat.com/browse/WINC-280

We're addressing WINC-280 as part of this PR as the node can be in unready state
and we're still counting it towards the windowsVM that was created. Only the
nodes that are in ready state should be accounted for when the operator
restarts