[wmco] Configure CNI
This commit run the wmcb's configure-cni command on the windows
nodes. It also transfers the files required for the configuration to the
windows VM. cni.conf file has been added in internal directory to be
transferred to the payload directory. In addition, added the east-west
networking tests to test connection between linux, windows pods.