[ci] Don't try to SSH into Windows VMs in tests
This commit stops the e2e tests from attempting to SSH into the spun up
Windows VMs. This is being done as once the WMCO is being run as a pod
in CI, the test pod will not have the ability to SSH into the VMs. This
is due to the ingress rules of the VMs not allowing access to port 22 to
anyone but internal cluster IPs, and the creator of the VM, which is
soon to be the WMCO pod.

The WindowsVM array was removed from the test framework and tests that
required it were switched to using node objects instead. This required
some changes throughout the networking tests, and additional retries
needed to be added to get around pre-downloading the Windows server 2019
container image.