Run kube-proxy on new Windows nodes
This commit adds running kube-proxy as a Windows service as part of the
procedure of configuring new Windows nodes. This enables communication
with workloads on the Windows node from outside the cluster.

The timeout for the e2e tests was extended as the new test added pushed
the test runtime past the previous timeout.