Bug 1840837: Get reconcile error on wmco reboot
This PR fixes the wmco reboot error that occurs due to
a stale resource read of the object from client cache for a
subsequent update. This PR replaces the default client
with a non-default-client that does not read from cache and
only updates the status to reconciling when we scale up or down.