[wmco] Update wget-ignore-cert.ps1 to run over ssh
cmdlets like wget may attempt to read the console mode of the standard
input or output streams if it is attached to the console causing
`Access is denied 0x5 occurred while reading the console output buffer`.
This can be avoided by setting them explicitly to null.

`$ProgressPreference = "SilentlyContinue"` is needed to prevent the
progress meter from accessing the console.

This is needed to run this script over SSH as it ends up being attached
to a console.