[wmco] Replace Run calls with RunOverSSH
In preparation for the move to using Machine API, replace all vm.Run()
calls with vm.RunOverSSH(). The main difference is that stdout and
stderr outputs are combined so we can't inspect  stderr as an error
indicator. Instead we have to depend on vm.RunOverSSH() returning error
on remote execution.