[wmco] Delete Machines with mismatched version annotations
This commit makes sure that machine controller handles changes in version
annotation on Windows nodes. When a node event occures, the machine
associated with that node will be reconciled. The controller deletes the
machines that results in drain and deletion of the node.This PR also
increases the value of nodeCreation in the e2e tests, to take into
consideration time required to delete the node and recreate it.

Co-authored-by: Vaishnavi Hire <vhire@redhat.com>