[ci] Dynamically choose OS and container images
This commit allows the testing to happen on both Windows Server 2019 and Windows
Server 1909 images. Please note the custom VXLAN port changing is available on
1903 and above images and 1909 is supported with Kube 1.19. However Windows
Server 2019 is still not having the change. So, this is a stop gap solution for
testing on both Windows Server and Windows 1909. Once we have  clarity on which
OS our customers are using, we will change the images in our test suite.