[wmco] Use downstream Kubernetes node binaries
This commit switches to using the kubelet and kube-proxy from
openshift/kubernetes. This is being done to pick up changes allowing
Windows pods to be scheduled outside of the default namespace.

This commit also includes changes to restructure the dockerfile to be
more consistant with the changes I have made, and pin cloned openshift
repositories to their 4.6 release branches.