[wmco] Watch cloud-private-key secret
This commit is stopping the mounting of the cloud-private-key secret,
and instead fetching it within the Windows Machine and secret
controllers.

This has two main benefits. Firstly, this allows the
operator to be deployed without the cloud-private-key existing. This
will allow the operator to be properly deployed via the OpenShift
console, allowing for a much better UX. Machines will not be configured
until the private key secret is created, as they will not enter the
provisoned state until the userData secret is created by WMCO.
Secondly, this commit allows for our operator to update the userData
secret immediately after a change is made to the private key.

This commit made the private key secret the primary resource of the
secret controller. This is more fitting than the userData being the
primary resource, as we create the userData based off the contents of
the private key. The private key is the source of truth and therefore
should be the primary resource.

Changes were also made to the CSV in order to remove the mounting of
the private key.

This commit only causes future Windows machines to be configured with
a new private key when it is specified, and does not deal with Windows
machines which were configured with a previous private key. That needs
to be handled in a future commit.