[wmco] Add version annotation after network config
This commit makes it so that WMCO waits to apply the node version
annotation until after network configuration has completed. This allows
for the catching of partially configured machines, machines that were
turned into nodes, but did not have all configuration steps completed.

This commit then makes it so that if the Windows machine configured
into a node but is missing the version annotation, WMCO attempts to
configure the machine once more.

This solves: https://bugzilla.redhat.com/show_bug.cgi?id=1875528