[wmco] Don't copy files again
Ignore transferring files, if they are already present. This is with assumption
that Windows VMs configuration is immutable for a given version of operator.
To make it clear, if we want a newer version of file that we want to transfer,
WMCO team should cut a newer version of operator. The files are tightly coupled
with the operator. Remove this when we do in place upgrades

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1875530