[wmco] Introduce vSphere changes
In case of vSphere:
- userData secret is not needed
- hostname of the windows VM that WMCO configures needs to be changed

Some more background:
MAO is responsible for changing hostname and applying userData scripts
for all linux nodes in OpenShift clusters however it uses ignition which
is RHCOS specific and cannot be used on Windows hosts. So, we are making
WMCO to change hostname of the Windows VM it configures. Since the userData
cannot be configured, it'll be a manual step during the Windows VM golden
image preparation.