[wmco] Update Node deletion during upgrade
Previously, during an upgrade, wmco defined minHealthy count. This count was
taken into consideration while deleting the nodes with mismatched version.
This commit updates the deletion process to take into account maxUnhealthy
count. This count ensures that we have only maxUnhealthy numbers of nodes in
that are not ready to serve during upgrades. For the first release this
value defaults to one.