[wmco] Remove affinity from north-south network test
The node affinity is not required for north-south test and removing that allows
us to deploy the windows webserver on any of the Windows Nodes during upgrade.
The name for the deployment no longer includes the machine ID, since it is not
tied to a particular machine after the removal of node affinity. This commit also
includes removing the deletion of Windows webserver, so that it can be used
during the upgrade test suite