[wmco] Run windows_exporter as a Windows service
This change was made to transfer the windows_exporter executable to the Windows
VM and run it as a Windows service, exposing node-specific collected metrics at
an endpoint reachable from within the cluster.

The promu prometheus build utility tool and windows_exporter are built
from the versions checked out in their respective submodules.

The windows_exporter service is only configured if the executable exists
on the Windows VM.

An end-to-end test is added to ensure each Windows node correctly
collects and exposes node-specific metrics (at default port 9182), thereby
checking that the windows_exporter service is running properly on
all Windows nodes.