fix k8s version  being misreported
This PR fixes the bug where kubelet and kube-proxy were not
reporting correct version.
This was because of absence of git metadata which is required for
setting this version.

The bug fix includes copying the git metadata.
Also, added test to check the kubelet version in-case it breaks again.
We are just checking if the prefix has the major and minor version like `v1.19`
Consolidated the node metadata test and included kubelet version there

Also, simplified the way we build the promu binary, by using `go install .`