Ensure all transfered files have the correct hash
This commit makes it so that we check the sha256 checksum of files that
were previously transfered to Windows nodes. If the checksum does not
match what we expect it to, the correct file will be transfered over.

As part of the work done to make this possible, all the services that
are started during the machine configuration are stopped at the start of
reconciliation. Because `sc stop` is an asynchronous operation,
functionality to wait for the service to stop was added to our
stopService function. This is needed to deal with stopping services
which have dependent services that need to be fully stopped before the
parent service can be stopped.

As part of this commit the wellknownlocations package was renamed to
payload.