[ci] Increase Windows curler retries
This commit increases the amount of times the Windows curlers will
retry. This is being done as the test which tests that Windows
containers scheduled on different nodes can communicate through a
ClusterIP service, has been flaking on vSphere. Increasing the retry
count will fix this issue.

In addition the overall test suite timeout had to be increased in order
to not time out while making the additional retries.