[wmco] Delete Machines when SSH auth fails
This commit prevents an issue where a user changes the private key used
to configure Windows Machines, after unconfigured Machines have been
created with the original userdata secret. When those unconfigured
Machines are reconciled they will never be able to be SSH'd into and
will infinitely be requeued.

This issue is being fixed by deleting Machines that have had an
authentication error so that they can be re-provisioned with userdata
corresponding with the new private key.

The test timeout was increased to allow for Machine recreation in the
test suite.