[wmco] Add pubkey annotation to Windows nodes
This commit adds a new annotation, 'windowsmachineconfig.openshift.io/pubkey'
to configured Windows nodes. The value of this annotation is the sha256
checksum of the public key associated with the private key used to
configure the corresponding VM.

If the private key used to configure nodes changes, Machines configured
with the old key will be deleted, so that they can be replaced with new
Machines.

(cherry picked from commit 9e5ea5b519eb85a10325f31f9091832d68caac39)