[vendor] Bump Kubernetes libraries to 1.20.0
This commit bumps all Kubernetes libraries being imported to Kubernetes
1.20.0.
This commit was done by changing changing the versions of the imports in
go.mod, and then running `go mod tidy` and `go mod vendor`.

The openshift/api and openshift/client-go libraries had to be updated as
well. This was done by replacing the libraries version with correct
commit, and then running `go mod tidy` and `go mod vendor`.