[ci] Use subtests in Node creation test suite
This commit reorganizes the creation test suite in order to be more
straightforward about what is occuring within the test. This is not
being done in a table driven test format, as the two test cases that we
care about have very little overlap. I find that doing it this way makes
the test much more readable.