Explicitly set GOBIN path for promu installation
This commit explicitly sets the GOBIN path
to specify where the promu binary should be
installed, promu installed in the windows_exporter
directory is then utilized for its installation.
This change was necessary because cachito changes
the GOPATH during image build and if the path is
not specified promu is installed to some location
from which it is not accessible to build
windows_exporter executable.