[ci] Introduce retries for WMCO deployment
This commit makes it so the e2e hack script retries multiple times in
case there is an transient issue when deploying the WMCO. Recently we
have been seeing an issue with Azure clusters going through an etcd
leader election soon after install. This change is being made as a
reaction to that, so that we retry WMCO deployment instead of having to
retry the entire test, including cluster provision.