[wmco] Remove constant for metrics resources name
This commit removes the constant used as name for  metrics resources.
The name for metrics resources is set by 'metrics' package of operator-sdk
and is equivalent to the name of the metrics service. The changes include
retrieving the name set by the 'metrics' package and using it for accessing
the resources.

(cherry picked from commit 831e42db1a24264fd28bb6f38e780e576ea3d6ce)