[test] Introduce CloudProvider.GetType()
And use it to only run the "Changing the private key ensures all Windows
nodes use the new private key" test on AWS. This cannot be run against
a vSphere cluster because this random key is not part of the vSphere
template image. Moreover this test is platform agnostic so is not needed
to be run for every supported platform.

(cherry picked from commit 34cb58f1828825e54af9381730a3b00a28fe1453)