[wmco] Reorganize pkg directory for OSDK 1.0
This commit reorganizes the pkg directory to be more like what is
used in projects created by operator-sdk v1.0+. In 1.0, projects are
created with a `controllers/` directory with each controller existing
flat in that directory. All controllers are part of the same
`controllers` package.

This is being done in accordance with the directions given here:
https://master.sdk.operatorframework.io/docs/building-operators/golang/migration/#what-was-changed

(cherry picked from commit 81610c5f8e0f623164874847042e1159982250f2)