[wmco] Use updated controller-runtime libraries
This commit makes it so that the controller-runtime libraries being
used by WMCO line up with operator-sdk version 1.4+. With this came
some major changes. All of these changes were made in accordance with
the migration guide https://master.sdk.operatorframework.io/docs/building-operators/golang/migration/

Notable architecture changes:

* The `Add()` functions for each controller were removed and replaced
with SetupWithManager() methods which are called in the main function.
* The `EnqueueRequestsFromMapFunc` way to queue objects was changed from
accepting an interface to deal with mapping, to a function. The mapping
functions were made a method of each reconciler.
* Linux node annotation cleanup now uses a kubernetes clientset instead
of the manager's client. This needs to be done as the manager's
client is not usable until it's `Start()` function is called. At that
point there is no way to call the cleanup function. This function was
made public as the kubernetes clientset is not needed anywhere else in
the secret reconciler, so I did not want to make the clientset part of
the SecretReconciler struct.

(cherry picked from commit 77a066b22009d098c6a225bb19772c7978652cb3)