Move controller loggers into Reconcilers
The loggers for each controller were moved into their respective
reconciler struct. This was done in preperation of the controllers
being moved into the same package.
The loggers for each controller shared the same name and needed to be
made unique in order to do the future move.
Having these loggers be part of each reconciler solves the
uniqueness problem, and is the way newly generated controllers handle
logging in OSDK1.x projects.
In addition the relevant structs in the nodeconfig and windows packages
had loggers added to them. This makes it so that we are using the same
pattern throughout the code