[wmco] Disable IPv6DualStack in kube-proxy
According to @danwinship, it appears that the windows kube-proxy startup
code is broken and incorrectly assumes that you won't enable the feature
gate unless you actually have a dual-stack configuration.

In the long term, we want to absorb NewDualStackProxier from
pkg/proxy/winkernel/proxier.go into cmd/kube-proxy/app/server_windows.go
, and implement "if the dual-stack feature gate is set (and we're using
the kernel proxier), then try to create both an IPv4 proxier and an IPv6
proxier, and if that succeeds then combine them together into a
metaproxier, but if one succeeds and one fails then fall back to a
single-stack using the successfully-created proxier".

In the short term, we are disabling IPv6DualStack in kube-proxy.