[build] Change Makefile to work for our repo
The goal of this commit was to assimilate the new targets required by
Operator SDK 1.5.1, while also maintaining compatibility with the
make targets used in CI.

The Golang image being used in CI sets the environment variable
`VERSION`. Because the Makefile is designed around being able to
override the default values with set environment variables,
theres an issue deploying the operator with packagemanifests.
In the Makefile $VERSION was replaced with $WMCO_VERSION to get around
this.