[deploy] Generate bundle folder and Dockerfile
This was generated by running `make bundle`.
After generation the Dockerfile had to have the Red Hat Operators labels
manually added to it.
Additionally the `:latest` portion of the image was manually removed in
order to maintain CI functionality. This will need to be removed
manually in the future whenever generating new bundles.