[wmco] Improve isValidMachine()
Introduce getInternalIPAddress() that returns the internal IPv4 address
of a Machine and use it in all the call sites. This has the good side
effect of improving isValidMachine() as it was not doing all the
necessary checks for IP addresses in Machine objects.

Remove the unit test that was checking for Hostname as a valid IP
address as we don't support that.