[wmco] Enable adding BYOH instances
This commit introduces the ability for Windows instances not created
through a Machine object to be configured by WMCO. This is part of
the Bring Your Own Host (BYOH) initiative, with the overall goal of
allowing pre-existing Windows instances to be joined to an OpenShift
cluster.

The functionality introduced within this commit is only the ability to
join Windows instances which fufill the listed requirements to a
cluster. This commit does not include the ability to remove or upgrade
the kubernetes binaries on said instance. Additionally, the CSRs for
the Node must be approved manually.

As part of this commit the `-n` option for hack/run-ci-e2e-test.sh was
taken out and replaced with `-c` and `-m` to indicate the instances to
be configured as nodes through the ConfigMap controller and Machine
controller, respectively.

One test was removed, checking that Machines without the Windows label
are not configured. This is still being implicitly tested, as to test
BYOH, Machines without the Windows label are expected to be configured
by the ConfigMap controller. Only the ConfigMap controller applies the
BYOH label to nodes, so if the Machine is configured by the Machine
controller, the creation tests will fail due to the missing annotation.