[wmco] Generalize NodeConfig for all Windows VMs
This commit is being done in an effort to generalize nodeconfig to work
for VMs not associated with a Machine object. Nodeconfig should not care
about Machine objects, instead should work with a VM provisioned in any
way.

The specific changes made in this commit are as follows:

* NewNodeConfig now accepts a hostname, instead of determining if the VM
should have its hostname changed based on it being a vSphere Machine.
A hostname is optional and will only cause an action if it is not an
empty string.

* NodeConfig no longer uses an instanceID when configuring VMs.
An instanceID is only available to WMCO when the VM is associated with
a Machine. This parameter was only being used as an identifier for
logging, and as a way to find nodes associated with VMs. In
order to allow for more general configuration, so that BYOH is
possible, the usage of instanceID to match nodes with VMs is now done
with IP address matching.

* A username is set by the caller of NewNodeConfig. This allows for the
usage of any user, not just those that are used when using Machine API
to provision VMs.
This also allows for the removal of knowledge of the `platform` of
nodeconfig, as that was only being used to determine what user should
the VM be configured through. Along with that, The addWorkerNode()
method of WindowsMachineReconciler, had its `platform` argument
removed, as thatinformation is already present in the object.