[wmco] Use controller runtime client in nodeconfig
Replaces usage of the client-go clientset in a nodeconfig object with
the client.Client interface from controller runtime. This allows the
usage of the cache client, which is preferable and possible to use as
nodeconfig only performs API operations on Nodes. The controller manager
also provides a non-cached client which fufills the client.Client
interface, which can be used in the future if needed.