[hack] Explicitly apply monitoring label
This commit explicitly applies the monitoring label
to be used when deploying the operator through OLM using
the hack/olm.sh script. This change needs to be made to
align with the changes made as part of OSDK migration.
Prior to the migration this label was applied through the
deploy/namespace.yaml which has now been replaced by a
section of config/manager/manager.yaml.