[test] Add label to each Machine created by e2e
Allows selection of all Machines and MachineSets created by the e2e for
cleanup, fixing an issue where only aws Machines were being cleaned up
in the deletion suite.

This commit paves the way for testing for BYOH as
waitForWindowsMachines() needs to be adapted to wait for Machines
without the `Windows` label. With this new label we will be able to
find those Machines easily by selecting Machines with the e2e label
and without the Windows label.

Additionally, the AWS provider wasnt using the constants the rest of the
providers were using for labels, so that was made consistent.