[test] Add CSR approval for vSphere
This is needed as the needed conditions for vSphere node CSRs approval
by the cluster-machine-approver are not being met as the BYOH vSphere
VM's hostname is not the same as the machine name.
This needs to be done for both the bootstrapping and the node CSR. This
is why the check is being made both at the beginning of the
waitForWindowsNodes function, as well as after we have determined the
node to be fully created.

Based on functionality previously in the operator, introduced in
f46ddec99d6f3fc0664403544764b136a5c6db7f.

This should be removed when the operator has CSR approval functionality
built in.