Ensure all Win services are running
This commit adds test code which is capable of determining the list of
Windows services on a Node, as well as their current state. A test was
added to creation suite ensuring that all the expected services are in
the running state.
In order to allow for SSHing from a pod to be possible, the private key
create and delete functions now create/delete the key in both the
operator namespace and the wmco-test namespace.

This commit enables future work to test that services are not present
after a BYOH node has been deconfigured.