Configure partially configured BYOH nodes
This commit ensures that BYOH nodes that were partially configured are
properly dealt with. The reconfiguration suite was expanded to include
reconfiguration of one Machine sourced node, and one BYOH node.

Before this commit, if an instance was far enough along in the
configuration process that a node object was created for it, and then
the configuration failed, the instance would be left in a partially
configured and unusable state, unless the node was manually deleted.