[wmco] BYOH node removal
This commit enables the removal of BYOH nodes from a cluster. A new
username annotation has been added to BYOH nodes in order to help
accomplish this.

Instances that are removed from the windows-instances ConfigMap will be
cordoned, drained, and deconfigured, returning them to the state they
were in before WMCO configured them.