Add node information to instance struct
This commit extends the instance struct by adding a pointer to the,
instance's associated node. This was done in order to allow functions
that are passed the instance to be able to determine the current
state of the instance in reference to the cluster.

The idea for this came from looking at the Machine object's NodeRef.
The instance object serves the same purpose as a Machine, but is more
general, in that it can describe both a BYOH instance, as well as a
Machine backed instance.

I considered using an ObjectReference, like Machines do with NodeRef,
but that does not make sense for our use case. An ObjectReference makes
sense in the case that we are storing the instance objects on the
cluster, such as in a CRD status, but since we are always going to have
the current state of the Nodes when reconciling, using the Node object
directly is the best choice here.