Upgrade nodes with older version annotation
Deconfigures and then configures any instances associated with a node
that has an mismatched WMCO version annotation. This currently only
applies to BYOH nodes, however this logic was placed in the
InstanceReconciler methods, and not the ConfigMapReconciler so that the
code can be shared between Machine and BYOH nodes in the future.