[wmco] Ensure APIServerInternal IP address at configuration
This commit adds the ensureAPIServerInternal() function to the windows
package, in order to ensure the IP address of the APIServerInternal
while configuring the VM.

If the APIServerInternal is available with an IP address, it's a no-op.

Otherwise, while no valid IP address resolution for APIServerInternal, the
APIServerInternal IP address is discovered from the cluster infrastructure,
specifically from the platform status, and included as a DNS alias entry
in the Windows hosts file.

Supported platform types: vsphere