[docs] Update vSphere documentation
This commit aligns the golden image documentation for
vSphere (vsphere-golden-image.md) with the proposed process
using Packer, where a ready to use virtual machine template
is produced.

Added guidelines to update the `MyPassword` placeholder in
the documentation.

Removed LocalAccount section in unattend.xml, no need to override the
current Administrator account details.

Disabled AutoLogon in unattend.xml, to avoid the security issue of
leaving an open terminal at boot.

There is a known issue, where the kubelet is not able to join a node if the
machine name is invalid, for example contains uppercase letters. To work around
this, the computer name is fixed to `winhost` in the unattend.xml file,
following the Kubernetes' names specification.
See https://kubernetes.io/docs/concepts/overview/working-with-objects/names

Co-authored-by: aravindhp <aravindh@redhat.com>