[wmco] Patch node when adjusting annotations
This commit uses a different API call to add/remove node annotations,
opting to Patch instead of Update. Update calls can result in failures
as the node object can be a stale reference (if some other entity changes
the Node while WMCO is configuring for example). A patch is best used when
modifying parts of a resource, which fits the node annotation use case;
this is why a patch was favored over a Get->Update loop.