[wmco] Combine annotations and metadata pkgs
Moves everything that was present in the annotations package into the
metadata package. This is being done as we will need to use the
functions present in the annotations package for labels as well, and it
makes more sense to generalize the functions considering the
similarities between annotations and labels.