[wmco] Refactor annotation patch creation
Refactors the patch creation in order to enable better unit testing.
Specifically, this commit makes it so that metadata.generatePatch
no longer returns the marshalled patches, but instead returns the slice
of patches it generated, allowing those to be marshalled at the call
sites, metadata.GenerateAddPatch() and metadata.GenerateRemovePatch().

This commit is important for two reasons. Firstly, using
assert.elementsMatch to test that a byte slice is correct does not seem
like the right thing to do. This seems as if it just tests that the
chars are all the same, so the patchdata could be made out of order and
this unit test would not catch that. Secondly this allows for more
complex testing, which will be required in a future commit where
generatePatch() handles both annotations and labels.

Additionally, since the tests had to be changed, I changed one case to
exercise the escaping functionality, something that was being missed by
the tests.